refactor: move module specific code to separate module in frontend-wasm
