+++ frontend-wasm/src/module/func_translator.rs
@@ -8,9 +8,9 @@
 
 use crate::code_translator::translate_operator;
 use crate::error::WasmResult;
-use crate::func_translation_state::FuncTranslationState;
-use crate::function_builder_ext::{FunctionBuilderContext, FunctionBuilderExt};
-use crate::module_env::ModuleInfo;
+use crate::module::environ::ModuleInfo;
+use crate::module::func_translation_state::FuncTranslationState;
+use crate::module::function_builder_ext::{FunctionBuilderContext, FunctionBuilderExt};
 use crate::ssa::Variable;
 use crate::translation_utils::emit_zero;
 use crate::wasm_types::valtype_to_type;
