+++ frontend-wasm/src/module/function_builder_ext.rs
