+++ frontend-wasm/src/wasm_types.rs
@@ -8,7 +8,7 @@ use miden_hir_type::Type;
 
 use crate::error::WasmError;
 use crate::error::WasmResult;
-use crate::module_env::ModuleInfo;
+use crate::module::environ::ModuleInfo;
 use crate::unsupported_diag;
 
 /// Index type of a function (imported or defined) inside the WebAssembly module.
